Input box accepts URL to find instagram post and display on site
Input box updates state instagramUrl with input text, and then onSubmit,
we parse the instagramUrl for the pictureId, which we use to query
the instagram API for the media file. Display this media file using
an img tag. When I finished this I was trying to figure out how to add
a download button that would allow you to download the photo once you
visually confirmed that the photo we retrieved was the correct one.
Haven't figured that out yet.